requirements.txt
fastapi
uvicorn
numpy
Pillow  # PIL fork
tensorflow
python-multipart
slowapi